+++ src/speaker.py
@@ -1,0 +1,55 @@
+from wave import WAVE_FORMAT_PCM
+import pyttsx3
+import wave
+import pyaudio
+
+engine = pyttsx3.init('sapi5')
+
+# Set Rate
+engine.setProperty('rate', 190)
+
+# Set Volume
+engine.setProperty('volume', 1.0)
+
+# Set Voice (Female)
+voices = engine.getProperty('voices')
+engine.setProperty('voice', voices[0].id)
+
+def speak(text):
+    """Used to speak whatever text is passed to it"""
+
+    engine.say(text)
+    engine.runAndWait()
+
+def play_sound (file_name):
+    chunk = 1024  
+
+    #open a wav format music  
+    f = wave.open(file_name,"rb")  
+    #instantiate PyAudio  
+    p = pyaudio.PyAudio()  
+    #open stream  
+    stream = p.open(format = p.get_format_from_width(f.getsampwidth()),  
+                    channels = f.getnchannels(),  
+                    rate = f.getframerate(),  
+                    output = True)  
+    #read data  
+    data = f.readframes(chunk)  
+
+    #play stream  
+    while data:  
+        stream.write(data)  
+        data = f.readframes(chunk)  
+
+    #stop stream  
+    stream.stop_stream()  
+    stream.close()  
+
+    #close PyAudio  
+    p.terminate()
+    
+def main():
+    print ("No Default Functionality-Import From An Existing script.")   
+    
+if __name__ == "__main__":
+    main()
